i386/sev: Support measured direct -kernel boot on SNP
This is an RFC patch based on AMD's RFC snp-v3 tree [1].

In SNP, the hashes page is not included in the ranges to pre-validate
that appear in the SNP metadata published by OVMF (See [2] for proposed
OVMF patches).

Therefore, if the user enabled kernel hashes (for measured direct boot),
we should fill hashes table and encrypt the page.  Note that in SNP
(unlike SEV and SEV-ES) the measurements is done in whole 4KB pages.
Therefore we zero the whole page that includes the hashes table, and
fill in the kernel hashes area in that page, and then encrypt the whole
page.  The rest of the page is reserved for SEV launch secrets which are
not usable anyway on SNP.

If the user disabled kernel hashes, QEMU pre-validates the page as a
zero page.

Note that the base branch [1] doesn't yet include the kernel-hashes
flag.

[1] https://github.com/AMDESE/qemu/tree/snp-v3
[2] https://edk2.groups.io/g/devel/message/88137

Signed-off-by: Dov Murik <dovmurik@linux.ibm.com>